chore(react): merge vitest.config.ts into vite.config.ts
Consolidate the split test config across the React packages into a single
`vite.config.ts` (matching the approach adopted in packages/svelte/
ds-app-launchpad, per the discussion on #693). Each package's `test` block is
folded verbatim onto the base `defineConfig` (now imported from `vitest/config`),
the standalone `vitest.config.ts` is removed, and the tsconfig `include` drops
its now-dead reference.

Pure refactor — the jsdom setup is unchanged: same environments, includes,
setupFiles, and (for ds-global / hooks) the two-project client+ssr split. The
Playwright/browser-mode migration remains the separate deferred item (#729).

Verified per package: the same test set runs against the unified config (tests
green), tsc has no new config-caused errors, biome clean. Covers ds-app,
ds-app-anbox, ds-app-landscape, ds-app-launchpad, ds-app-lxd, ds-app-portal,
ds-global, ds-global-form, and hooks.
