+++ packages/react/ds-global-form/tsconfig.json
@@ -21,7 +21,6 @@
     ".storybook/*.ts",
     ".storybook/*.tsx",
     "vite.config.ts",
-    "vitest.setup.ts",
-    "vitest.config.ts"
+    "vitest.setup.ts"
   ]
 }
